fix(ci): clear ruff lint and OpenMW collection ImportError
pytest-cov plus numpy 2.4+ re-inits _multiarray_umath (PEP 489) and
collection dies on every OpenMW test. scipy 1.16+ then trips
_CopyMode.IF_NEEDED on that reload. Pin numpy<2.3 and scipy<1.16.

Clear the 93 ruff findings that fail CI without weakening the gate.
Keep lastEventId as a Query alias and UnslothNotAvailable as an alias.

+++ OpenMW/openmw/openvault/vault/health_store.py
@@ -82,8 +82,8 @@ def _percentile(sorted_vals: list[float], pct: float) -> float | None:
     if len(sorted_vals) == 1:
         return sorted_vals[0]
     rank = (pct / 100.0) * (len(sorted_vals) - 1)
-    lo = int(math.floor(rank))
-    hi = int(math.ceil(rank))
+    lo = math.floor(rank)
+    hi = math.ceil(rank)
     if lo == hi:
         return sorted_vals[lo]
     frac = rank - lo
@@ -128,9 +128,7 @@ class HealthStore:
             )
             conn.commit()
 
-    def _latest_row(
-        self, conn: sqlite3.Connection, key_id: str
-    ) -> tuple[float, str] | None:
+    def _latest_row(self, conn: sqlite3.Connection, key_id: str) -> tuple[float, str] | None:
         row = conn.execute(
             """
             SELECT checked_at, status FROM precheck_history
